+++ tests/e2e/singlecard/core/test_ascend_scheduler.py
