+++ tests/e2e/singlecard/ops/test_rotary_embedding.py
