+++ tests/e2e/singlecard/test_offline_inference.py
