+++ tests/e2e/singlecard/test_prompt_embedding.py
